Nonworking code for future expansion.
There's new code on the test suites for the asm, file 'reader.c' that's
not actually test code. It's code I was writing while trying a top-down
design of the reader interface by "using" it to load a module.

Along with it, there's a new unfinished doc on the grammar, that shows
what grammar that reader is supposed to handle.

Works for nothing, but breaks nothing also. And I need to be able to
check it out on other computers later on. So it's here :)