Changed nomenclature scheme for types.
Previously I've been using the form <prefix>_<name>_t, where <prefix>
was either "n" or "ni" if the type was to be externally visible or not,
and name was the lower case name of the type with words separated by
underscores.

It turns out that, though unlikely, this naming scheme conflicts with an
extension avenue from the POSIX standard, that says that symbols ending
with _t may be used in the future to denote new POSIX types.

Considering a few questions I decided to fall back to a new nomenclature
scheme N<name>, where <name> is the name of the type in CamelCase,
Java-style.

The following reasons are behind this choice:

1. Types won't be visible from the exterior unless they're exposed in a
header file (unlike functions/variables, which'll be visible in the
object files symbol list), so I don't really need the different prefixes
(n or ni) for types.

2. It saves me a few characters in type names, which makes for a little
more room for code before I have to break lines.

3. I already have a bunch of code from nuvm-second-experiment where it
uses this pattern for type names, and it'll make it easier to reuse such
older code.