Reworking test assertions.
Creating assertions for tests was becoming too much of a chore when I
started to develop common assertions (like 'equals' for int64_t, etc).
Also, management of resources (mainly, disposing of resources consumed
during the execution of the test in case of failure) was impossible in
some cases.

A big refactor went into making these things easier and possible, which
lead me to the separation of the assertions into a generic ASSERT_CHECK
macro, that runs a check and returns its success status and error
message, if any, and a bunch of checkers, such as CK_EQ_I64 (checks if
two int64_t are equal) and others (found in test-suite.{h,c}.