Assembler overhaul.
The Assembler object (asm/assembler.c) has received major (untested)
code functionality. So far I'm thinking I'll only be able to test it the
way it's actually intended to be used: as a major facade with only one
or two entry points.

To accomodate these additions, a bunch of code was added to asm/reader.c
as well, and this part needs tests which can be reasonably done.