Improve Assembler with instructions support.
Preliminar support for instructions in the Assembler object
(src/asm/assembler.c) is added. Arithmetic and Relational opcodes are
already supported in the reading phase, with infrastructure that seems
to allow that only the reader (src/asm/reader.c) must be updated to
support new instructions.

I still need to check on how label resolution will work, but that's for
the generation phase, which is not under scrutiny yet.