Remove length function from istreams
NIStreams had a length function that made no sense when considering streams
should be generic. We can't always know the size of the stream so there
can't be a function to tell its length.

Since I was already there I also added a bit of error handling to the
creation of streams.