Factor out label functionality from assembler.
The label handling functionality of the assembler is now taken to its own
file, the Label Manager. It's still in the assembler, so the next step is to
take off the code from the assembler and delegate to the manager code.

I came to realize the assembler is getting too big because it's trying to do
too much. Next commits will be about trimming the assembler so that its only
responsibility is to receive a lexer and produce a representation of the
file. This representation can then be passed to a emitter whose work is to
write off this general representation on a more packed, binary,
representation using the output streams.