Make Label Manager stack allocable.
The Label Manager can now be allocated inside structs or in the stack. This
will help diminish our malloc footprint. One of the future goals is to get
rid of as much dynamic allocation as possible in the inner workings. I
believe the less dynamic allocation we have the less interference with the
garbage collector when we finally get to implement it.