Add the detach option to resizable array
The detach option makes it so that, upon destruction, the backing array is
not freed. It's supposed to be used in situations where you first use the
resizable array to build an array, then use the backing array it built to
pass along to other objects without the rest of the structure.