Simplify assembler interface.
The assembler object is only needed for maintaining the internal state
between calls to the static functions inside the module after calling the
read_from_istream function. That being so, I'm pulling the type definitions
and new/destroy functions inside the .c file and the read_from_istream
function receives the NIStream only.