Remove instructions unnecessary heap allocation.
During the work of the assembler, we clone instructions to keep ownership
over them. This cloning is unnecessary since on the very next lines we
destroy the original we're cloning from. Use copied elements and stack
allocation whenever possible to minimize heap allocation. The number of
instruction copying seems to remain the same, only now it's implicit, in
code I don't have to write.