Make char/string constants record their size.
This makes it easier when we get to emit the strings and character constants
because they'll serialize with their length with them. Since we already
compute it when we're parsing the constants data it only makes sese to save
them down there and bubble up to the emitter when it needs, instead of
recomputing it unnecessarily.