Simplify instructions register arguments.
Now each argument can only be of one kind of register -- local, global or
constant. This makes a few scenarios more complicated: since most operations
now only work on locals, operating on constants need to first load their
value into a local. On the other hand the code is greatly simplified, and I
believe this is more important for the purposes of this VM.

A couple new instructions were needed and have already been described in the
docs, but their implementation hasn't landed yet. It's minor work, though,
and I'll probablly finish it tomorrow.