Create a couple of instructions and rename one.
The 'move' instruction has now been renamed to 'move-in' ("in" being a
reference to it moving values from the outer range of locals -- above 255).
Two new instructions have also been created: "move-out" which is the counter
point to "move-in" and "const-ref", which does for constants what
"global-ref" does for globals.