Change jump instructions to use labels.
This was an oversight of mine. Part of the code was admitting there would be
instructions using labels to determine offsets, while the tests and
implementation were still treating them (jump and jump-if) as if they would
receive the signed integers describing the jump. It's unreasonable to expect
the programmer (and to a lesser extent, the compiler) to compute these
values unnecessarily, since we can easily handle them in the assembler.
Now both instructions take labels as arguments.