Separate instruction decoders from encoders.
This avoids a bunch of warnings on unused functions. The decoders are
probably going to be changed later on for performance reasons, but that
will only happen after I can profile the whole business. Profiling will
take a while, but I want these separated in a header file so it can be
easily inlined if the need arises.